libreria

+++ app/src/main/java/com/riobamba/geolam/modelo/Proceso.java
@@ -2,17 +2,17 @@ package com.riobamba.geolam.modelo;
 
 public class Proceso {
 
-    public double obtenerDistancia(Double latitudUsu, Double longitudUsu, Double latitudLug, Double longitudLug)
+    public double obtenerDistancia(Double latitudUsu, Double longitudUsu, Float latitudLug, Float longitudLug)
     {
-        Double latitud = convertirRadianes(latitudUsu) - convertirRadianes(latitudLug);
-        Double longitud =convertirRadianes(longitudUsu) -convertirRadianes(longitudLug);
+        double latitud = convertirRadianes(latitudUsu) - convertirRadianes((double) latitudLug);
+        double longitud =convertirRadianes(longitudUsu) -convertirRadianes((double)longitudLug);
 
-        Double a = Math.pow(Math.sin(latitud/2),2)  +
-                Math.cos(convertirRadianes(latitudLug))*
+        double a = Math.pow(Math.sin(latitud/2),2)  +
+                Math.cos(convertirRadianes((double)latitudLug))*
                 Math.cos(convertirRadianes(latitudUsu))*
                         Math.pow(Math.sin(longitud/2),2);
 
-        Double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
+        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
 
         return 6371*c;
